feature : Redis DB 서버 연동 완료 및 이메일 인증 구현 완
