--- user-service/src/main/java/com/petful/userservice/UserServiceApplication.java
@@ -1,13 +1,0 @@
-package com.petful.userservice;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class UserServiceApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(UserServiceApplication.class, args);
-    }
-
-}
