--- user-service/src/main/java/com/petful/userservice/domain/Role.java
@@ -1,7 +1,0 @@
-package com.petful.userservice.domain;
-
-public enum Role {
-    User,
-    Admin,
-    Advertiser
-}
