--- user-service/src/main/java/com/petful/userservice/dto/AuthResponse.java
@@ -1,17 +1,0 @@
-package com.petful.userservice.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-public class AuthResponse {
-    private String token;  // 로그인 시에만 제공
-    private String message;
-    private String email;
-    private String name;
-}
