--- user-service/src/main/java/com/petful/userservice/dto/EmailVerificationConfirmRequest.java
@@ -1,15 +1,0 @@
-package com.petful.userservice.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-public class EmailVerificationConfirmRequest {
-    private String email;
-    private String verificationCode;
-}
