--- user-service/src/main/java/com/petful/userservice/dto/EmailVerificationRequest.java
@@ -1,14 +1,0 @@
-package com.petful.userservice.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-public class EmailVerificationRequest {
-    private String email;
-}
