--- user-service/src/main/java/com/petful/userservice/dto/TokenValidationResponse.java
@@ -1,18 +1,0 @@
-package com.petful.userservice.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-public class TokenValidationResponse {
-    private boolean valid;
-    private String email;
-    private String name;
-    private String role;
-    private String message;
-}
