--- user-service/src/main/java/com/petful/userservice/repository/UserRepository.java
@@ -1,13 +1,0 @@
-package com.petful.userservice.repository;
-
-import com.petful.userservice.domain.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-
-@Repository
-public interface UserRepository extends JpaRepository<User, Long> {
-    Optional<User> findByEmail(String email);
-    boolean existsByEmail(String email);
-}
