+++ user-service/src/main/java/site/petful/userservice/domain/Role.java
@@ -1,0 +1,7 @@
+package site.petful.userservice.domain;
+
+public enum Role {
+    User,
+    Admin,
+    Advertiser
+}
