+++ user-service/src/main/java/site/petful/userservice/dto/EmailVerificationConfirmRequest.java
@@ -1,0 +1,15 @@
+package site.petful.userservice.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor
+public class EmailVerificationConfirmRequest {
+    private String email;
+    private String code;
+}
