+++ user-service/src/main/java/site/petful/userservice/dto/EmailVerificationRequest.java
@@ -1,0 +1,14 @@
+package site.petful.userservice.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor
+public class EmailVerificationRequest {
+    private String email;
+}
