+++ user-service/src/main/java/site/petful/userservice/dto/TokenValidationRequest.java
@@ -1,0 +1,14 @@
+package site.petful.userservice.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor
+public class TokenValidationRequest {
+    private String token;
+}
