+++ user-service/src/main/java/site/petful/userservice/dto/TokenValidationResponse.java
@@ -1,0 +1,18 @@
+package site.petful.userservice.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor
+public class TokenValidationResponse {
+    private boolean valid;
+    private String email;
+    private String name;
+    private String role;
+    private String message;
+}
