+++ user-service/src/main/java/site/petful/userservice/service/EmailService.java
@@ -1,0 +1,9 @@
+package site.petful.userservice.service;
+
+import site.petful.userservice.dto.EmailVerificationConfirmRequest;
+import site.petful.userservice.dto.EmailVerificationRequest;
+
+public interface EmailService {
+    void sendVerificationEmail(EmailVerificationRequest request);
+    boolean verifyEmailCode(EmailVerificationConfirmRequest request);
+}
